employee_portal/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
email-validator==2.1.1
eventlet==0.36.1
requests>=2.31.0

